Flowmatrix(LIM)	R Documentation

Generates a flow matrix for an inverse (foodweb) problem

Description

Given a linear inverse model food web input list, generates a flow matrix, that contains the values of flows

Usage

Flowmatrix(lim, web=NULL)

Arguments

lim 	 a list that contains the linear inverse model specification, as generated by function Setup

web 	 the solved (food) web problem; if not specified, the model is solved first, using Ldei

Value

the flow matrix, containing the magnitude of the flows. The value on row i and column j is the flow *from* i and *to* j

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

plotweb from package diagram, which takes as input the flow matrix and plots the food web

Examples

Flowmatrix(LIMRigaAutumn)

[Package LIM version 1.1 Index]