LIM-package(LIM)	R Documentation

Linear Inverse Modelling

Description

functions that read and solve linear inverse problems (food web problems, linear programming problems, flux balance analysis). These problems find solutions to linear or quadratic functions: min or max (f(x)), where f(x) = ||Ax-b||^2 or f(x) = sum(ai*xi) subject to equality constraints Ex=f and inequality constraints Gx>=h. Uses package limSolve.

Details

Package: 	 LIM

Type: 	 Package

Version: 	 1.1-1

Date: 	 2008-07-17

License: 	 GNU Public License 2 or above

The model problem is formulated in text files in a way that is natural and comprehensible. LIM then converts this input into the required linear equality and inequality conditions, which can be solved either by least squares or by linear programming techniques. By letting an algorithm formulate the mathematics, it is simple to reformulate the model in case a parameter value changes, or a component is added or removed.

Three different types of problems are supported: flow networks, reaction networks and other (operations research) problems. The first two cases are based on mass balances of the components.

The package includes many examples

Author(s)

Karline Soetaert (Maintainer), Dick van Oevelen

See Also

Read, Setup for reading files and creating the model

Ldei, Lsei,Linp, Flowmatrix, Plotranges, Variables, Varranges, Xranges, Xsample.

Examples


## Not run: 
## show examples (see respective help pages for details)
example(Lsei)
example(LIMRigaSpring)
example(Ldei)
example(Xsample)
example(Varranges)

## run demos
demo("LIMexamples")

## open the directory with R sourcecode examples
browseURL(paste(system.file(package="LIM"), "/examples/Foodweb", sep=""))
browseURL(paste(system.file(package="LIM"), "/examples/LinearProg", sep=""))
browseURL(paste(system.file(package="LIM"), "/examples/Reactions", sep=""))

## show package vignette with tutorial about how to create input files
vignette("LIM")
browseURL(paste(system.file(package="LIM"), "/doc", sep=""))
## End(Not run)

[Package LIM version 1.1 Index]