LIMBrouageMudflat(LIM)	R Documentation

Linear inverse model specification for the Intertidal mudflat food web on the Atlantic coast of France

Description

Linear inverse model specification for the Intertidal mudflat food web on the Atlantic coast of France
as in Leguerrier et al., 2003.

The foodweb comprises 16 functional compartments and 3 external compartments, connected with 95 flows. Units of the flows are g C/m2/year

The linear inverse model LIMBrouageMudflat is generated from the file BrouageMudflat.input which can be found in subdirectory /examples/FoodWeb of the packages directory

In this subdirectory you will find many foodweb example input files
These files can be read using Read and their output processed by Setup which will produce a linear inverse problem specification similar to LIMBrouageMudflat

Usage

data(LIMBrouageMudflat)

Format

a list of matrices, vectors, names and values that specify the linear inverse model problem.

see the return value of Setup for more information about this list

A more complete description of this structures is in vignette("LIM")

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl> Dick van Oevelen<d.vanoevelen@nioo.knaw.nl>

References

Leguerrier, D., Niquil, N., Boileau, N., Rzeznik, J., Sauriau, P.G., Le Moine, O., Bacher, C., 2003. Numerical analysis of the food web of an intertidal mudflat ecosystem on the Atlantic coast of France. Marine Ecology Progress Series 246, 17-37.

See Also

LIMTakapoto, LIMRigaSummer and many others

Examples

Brouage <- Flowmatrix(LIMBrouageMudflat)
plotweb(Brouage,main="Brouage mudflat food web",sub="gC/m2/yr")
# Some ranges are infinite ->marked with "*
Plotranges(LIMBrouageMudflat,lab.cex=0.7, sub="*=unbounded",
               xlab="gC/m2/year",main="Brouage mudflat, Flowranges")
Plotranges(LIMBrouageMudflat,type="V",lab.cex=0.7, sub="*=unbounded",
               xlab="gC/m2/year",main="Brouage mudflat, Variable ranges")
               

[Package LIM version 1.1 Index]