LIMScheldtIntertidal(LIM)	R Documentation

Linear inverse model specification for the Schelde Intertidal flat food web

Description

Linear inverse model specification for the Westerschelde Intertidal flat food web in June
as in Van Oevelen et al. (2006).

The Westerschelde is a highly eutrophic estuary in the Netherlands. The food web model was created for the intertidal flat called the "Molenplaat", site 2. It is the basic input model.
The foodweb comprises 7 functional compartments and five external compartments, connected with 32 flows. Units of the flows are mg C/m2/day

The linear inverse model LIMScheldtIntertidal is generated from the file ScheldtIntertidal.input which can be found in subdirectory /examples/FoodWeb of the packages directory

In this subdirectory you will find many foodweb example input files
These files can be read using Read and their output processed by Setup which will produce a linear inverse problem specification similar to LIMScheldtIntertidal

Usage

data(LIMScheldtIntertidal)

Format

a list of matrices, vectors, names and values that specify the linear inverse model problem.

see the return value of Setup for more information about this list

A more complete description of this structures is in vignette("LIM")

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl> Dick van Oevelen<d.vanoevelen@nioo.knaw.nl>

References

Van Oevelen, D., Soetaert, K., Middelburg, J.J., Herman, P.M.J., Moodley, L., Hamels, I., Moens, T., Heip, C.H.R., 2006b. Carbon flows through a benthic food web: Integrating biomass, isotope and tracer data. J. Mar. Res. 64, 1-30.

See Also

LIMTakapoto, LIMRigaSummer and many others

Examples

ScheldtIntertidal <- Flowmatrix(LIMScheldtIntertidal)
plotweb(ScheldtIntertidal,main="Scheldt intertidal flat food web",
        sub="mgC/m2/day")
Plotranges(LIMScheldtIntertidal,lab.cex=0.7,
        main="Scheldt intertidal flat food web, Flowranges")
Plotranges(LIMScheldtIntertidal,type="V",lab.cex=0.7,
        main="Scheldt intertidal flat food web, Variable ranges")

[Package LIM version 1.1 Index]