Variables(LIM)	R Documentation

Generates the values of variables for a linear inverse (foodweb) problem

Description

Given an linear inverse model input list, generates the values of the inverse variables

Usage

Variables(lim, res, ...)

Arguments

lim 	 a list that contains the linear inverse model or linear programming specification, as generated by function setup.limfile

res 	 the solved linear inverse problem; if not specified, the model is solved first, using Lsei.lim<

... 	 extra parameters passed to function Lsei.lim

Value

the variables, a one-column data.frame

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

Varranges which estimates the ranges of variables

Examples

Variables(LIMRigaAutumn)

[Package LIM version 1.1 Index]