Varranges(LIM)	R Documentation

Generates ranges of the variables for a linear inverse problem

Description

Given an inverse input list, generates the minimal and maximal values of the variables (linear combinations of unknowns)

Usage

Varranges(lim,...)

Arguments

lim 	 a list that contains the linear inverse model or linear programming specification, as generated by function Setup

... 	 extra arguments passed to function varranges

Value

a 2-columned vector containing the minimum (column 1) and maximum (column 2) of each variable

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

Xranges which estimates the ranges of unknowns
Plotranges to plot the ranges

Examples

Varranges(LIMRigaAutumn)

[Package LIM version 1.1 Index]