Xranges(LIM)	R Documentation

Generates ranges of the unknowns of a linear inverse problem

Description

Given an inverse input list, generates the minimal and maximal values of the unknowns

Usage

Xranges(lim,...)

Arguments

lim 	 a list that contains the linear inverse model or linear programming specification, as generated by function setup.limfile

... 	 extra arguments passed to function xranges

Value

a 2-columned vector containing the minimum (column 1) and maximum (column 2) of each unknown

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

Varranges which estimates the ranges of inverse variables
Plotranges to plot the ranges

Examples

# ranges
  xr <- Xranges(LIMRigaAutumn)
  xlim <- range(xr)
  # parsimonious
        xpars <- Lsei(LIMRigaAutumn)$X
  dotchart(x=xpars,labels=rownames(xr),xlim=xlim,                                                        
           main="Riga Autumn ",
           sub="ranges and parsimonious solution",pch=16)
  cc <- 1:nrow(xr)
  segments(xr[,1],cc,xr[,2],cc)

[Package LIM version 1.1 Index]