Xsample(LIM)	R Documentation

Generates a random sample of the unknowns for a linear inverse problem

Description

Given an inverse input list, randomly samples the unknowns, using a Bayesian MCMC method

Usage

Xsample(lim,exact=NULL,...)

Arguments

lim 	 a list that contains the linear inverse model or linear programming specification, as generated by function Setup

exact 	 if not null, a vector containing the numbers of the equations to be solved exactly; if NULL, all equations are considered exact

... 	 extra parameters passed to function xsample

Details

For overdetermined LIM problems, the inverse equations may be split up in equations which have to be exactly met and other equations which have to be approximated.
exact is a vector with the exact equations

The default settings of xsample will often not do. For instance, the default consists of 3000 iterations (iter) and a jump length of jmp of 0.1. You may need to increase one of those to ensure that the entire solution space has been adequately sampled.

Value

a 2-columned vector containing the minimum (column 1) and maximum (column 2) of each unknown

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

Varranges which estimates the ranges of inverse variables
Plotranges to plot the ranges

Examples

# sample solution space
  xs <- Xsample(LIMRigaAutumn,iter=500,jmp=5)
  # remove flows that are invariable (sd=0)
  xs <- xs[,-which(sd(xs)==0)]
  #pairs plot
  pairs(xs,gap=0,pch=".",upper.panel=NULL)

[Package LIM version 1.1 Index]