Plotranges(LIM)	R Documentation

Plots minimum and maximum ranges. Takes as input either a lim list as generated by Setup or a set of vectors specifying the minimu maximum and the central value or a data.frame that contains min max and central values.

Description

Plots the minimum and maximum and central values

Usage


Plotranges.double(min, max, value=NULL, labels=NULL,log="",pch=16,
pch.col="black", line.col="gray", seg.col="black", xlim=NULL, 
main=NULL, xlab=NULL, ylab=NULL, lab.cex=1.0, ...)
Plotranges.lim(lim=NULL,labels=NULL,log="",pch=16,
pch.col="black", line.col="gray", seg.col="black", xlim=NULL,
main=NULL, xlab=NULL, ylab=NULL, lab.cex=1.0, ...)
Plotranges.character(file,...)
Plotranges(...)

Arguments

min 	 minimum value

max 	 maximum value

value 	 median or mean value

lim 	 a list that contains the linear inverse model specification, as generated by function Setup

file 	 name of the linear inverse input file

labels 	 names of each value

log 	 if = x: logarithmic scale for x-axis

pch 	 pch symbol used for mean value

pch.col 	 pch color for mean value

line.col 	 color for each variable, spanning x-axis

seg.col 	 color for variable range

xlim 	 limits on x-axis

main 	 main title

xlab 	 x-axis label

ylab 	 y-axis label

lab.cex 	 label expansion value

... 	 arguments passed to R-function "text" when writing labels

Value

Only when a lim list was inputted. A data frame with

min 	 the minimum

max 	 the maximum

values 	 the central value

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

[Package LIM version 1.1 Index]